fix(cardinal): submit all transactions to chain
