+++ cardinal/server/persona.go
@@ -1,15 +1,18 @@
 package server
 
 import (
-	"encoding/json"
+	"context"
 	"errors"
-	"io"
-	"net/http"
-
-	"github.com/invopop/jsonschema"
+	"fmt"
 	"pkg.world.dev/world-engine/cardinal/ecs"
 	"pkg.world.dev/world-engine/cardinal/ecs/transaction"
 	"pkg.world.dev/world-engine/sign"
+)
+
+const (
+	getSignerForPersonaStatusUnknown   = "unknown"
+	getSignerForPersonaStatusAvailable = "available"
+	getSignerForPersonaStatusAssigned  = "assigned"
 )
 
 // ReadPersonaSignerRequest is the desired request body for the read-persona-signer endpoint.
@@ -47,82 +50,41 @@ func (t *Handler) getPersonaSignerResponse(req *ReadPersonaSignerRequest) (*Read
 	return &res, nil
 }
 
-func (t *Handler) handleReadPersonaSigner(w http.ResponseWriter, r *http.Request) {
-	buf, err := io.ReadAll(r.Body)
-	if err != nil {
-		writeError(w, "unable to read body", err)
-		return
-	}
-
-	req, err := decode[ReadPersonaSignerRequest](buf)
-	if err != nil {
-		writeError(w, "unable to decode body", err)
-		return
-	}
-
-	res, err := t.getPersonaSignerResponse(&req)
-	if err != nil {
-		writeError(w, "read persona signer error", err)
-		return
-	}
-
-	resJson, err := json.Marshal(res)
-	if err != nil {
-		writeError(w, "unable to marshal response", err)
-		return
-	}
-
-	writeResult(w, resJson)
-}
-
-func (t *Handler) handleReadPersonaSignerSchema(w http.ResponseWriter, _ *http.Request) {
-	jsonSchema, err := json.Marshal(jsonschema.Reflect(new(ReadPersonaSignerRequest)))
-	if err != nil {
-		writeError(w, "unable to marshal response", err)
-		return
-	}
-
-	writeResult(w, jsonSchema)
-}
-
-func generateCreatePersonaResponseFromPayload(payload []byte, sp *sign.SignedPayload, tx transaction.ITransaction, world *ecs.World) (*TransactionReply, error) {
+func (t *Handler) generateCreatePersonaResponseFromPayload(payload []byte, sp *sign.SignedPayload, tx transaction.ITransaction) (*TransactionReply, error) {
 	txVal, err := tx.Decode(payload)
 	if err != nil {
 		return nil, errors.New("unable to decode transaction")
 	}
-	tick, txHash := world.AddTransaction(tx.ID(), txVal, sp)
-	res := TransactionReply{
-		TxHash: string(txHash),
-		Tick:   tick,
-	}
-	return &res, nil
+	return t.submitTransaction(txVal, tx, sp)
 }
 
-func (t *Handler) makeCreatePersonaHandler(tx transaction.ITransaction) http.HandlerFunc {
-	return func(writer http.ResponseWriter, request *http.Request) {
-		payload, sp, err := t.verifySignatureOfHTTPRequest(request, true)
-		if err != nil {
-			if errors.Is(err, ErrorInvalidSignature) {
-				writeUnauthorized(writer, err)
-				return
-			} else if errors.Is(err, ErrorSystemTransactionRequired) {
-				writeUnauthorized(writer, err)
-				return
-			}
-			writeError(writer, "unable to verify signature", err)
-			return
-		}
-		txReply, err := generateCreatePersonaResponseFromPayload(payload, sp, tx, t.w)
-		if err != nil {
-			writeError(writer, "", err)
-			return
-		}
-		res, err := json.Marshal(txReply)
-		if err != nil {
-			writeError(writer, "unable to marshal response", err)
-			return
-		}
+// submitTransaction submits a transaction to the game world, as well as the blockchain.
+func (t *Handler) submitTransaction(txVal any, tx transaction.ITransaction, sp *sign.SignedPayload) (*TransactionReply, error) {
 
-		writeResult(writer, res)
+	submitTx := func() *TransactionReply {
+		tick, txHash := t.w.AddTransaction(tx.ID(), txVal, sp)
+
+		return &TransactionReply{
+			TxHash: string(txHash),
+			Tick:   tick,
+		}
+	}
+
+	// check if we have an adapter
+	if t.adapter != nil {
+		// if the world is recovering via adapter, we shouldn't accept transactions.
+		if t.w.IsRecovering() {
+			return nil, errors.New("unable to submit transactions: game world is recovering state")
+		} else {
+			txReply := submitTx()
+			err := t.adapter.Submit(context.Background(), sp, uint64(tx.ID()), txReply.Tick)
+			if err != nil {
+				return nil, fmt.Errorf("error submitting transaction to blockchain: %w", err)
+			}
+			return txReply, nil
+		}
+	} else {
+		// if there is no adapter, then we can just put the tx in the queue.
+		return submitTx(), nil
 	}
 }
